Added support for MS2 and other ion mobility types
Added compatibility with MS2 and ion mobility data from instruments aside from timsTOF. Added more metadata that that is used, such as retention time, pixel_size, and max_dimension in addition to those directly relating to MS2. Made inputting compression types easier by allowing for string inputs. Changed the _spectrum and _spectrum_Mobility objects to be a dictionary to allow for easier addition of other variables. The ability to use just the variables needed for previous versions was maintained.